requirements.txt
numpy
trimesh[easy]
Pillow
httpx
pyglet<2